Start migrating project to Vue3 (WIP)
