--- babel.config.js
@@ -1,10 +1,0 @@
-module.exports = {
-    presets: [
-        '@vue/app',
-    ],
-    plugins: [['component', {
-        libraryName: 'element-ui',
-        styleLibraryName: 'theme-chalk',
-    },
-    ]],
-};
