+++ index.html
@@ -1,0 +1,45 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width,initial-scale=1.0,user-scalable=no">
+    <meta name="description" content="Share your location with TICE and see in real-time where your friends are and where they are heading. Secure by using end-to-end encryption.">
+    <meta name="author" content="TICE">
+    <meta name="apple-itunes-app" content="app-id=1494324936">
+
+    <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png?v=9B9XQqKnx0">
+    <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png?v=9B9XQqKnx0">
+    <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png?v=9B9XQqKnx0">
+    <link rel="manifest" href="/site.webmanifest?v=9B9XQqKnx0">
+    <link rel="mask-icon" href="/safari-pinned-tab.svg?v=9B9XQqKnx0" color="#2980b9">
+    <link rel="shortcut icon" href="/favicon.ico?v=9B9XQqKnx0">
+    <meta name="apple-mobile-web-app-title" content="TICE">
+    <meta name="application-name" content="TICE">
+    <meta name="msapplication-TileColor" content="#2980b9">
+    <meta name="theme-color" content="#ffffff">
+
+    <title>TICE – Live & Secure Location Sharing</title>
+
+    <meta property="og:title" content="Track my live location with the secure TICE app" />
+    <meta property="og:type" content="website" />
+    <meta property="og:url" content="https://ticeapp.com" />
+    <meta property="og:image" content="/OpenGraphTICEimg.png" />
+    <meta property="og:description" content="Share your location with me, so we find each other easily and feel safe. Meet up easily with TICE." />
+
+    <link href="https://api.tiles.mapbox.com/mapbox-gl-js/v1.11.0/mapbox-gl.css" rel="stylesheet"/>
+
+    <script type="text/javascript">
+      document.addEventListener('gesturestart', function (e) {
+        e.preventDefault();
+      });
+    </script>
+  </head>
+  <body>
+    <noscript>
+      <strong>We're sorry but TICE doesn't work without JavaScript enabled. Please enable it to continue.</strong>
+    </noscript>
+    <div id="app"></div>
+    <script type="module" src="/src/main.ts"></script>
+  </body>
+</html>
