+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "tice-web",
-  "version": "2.1.0",
+  "name": "tice-web-vue3",
+  "version": "0.0.0",
   "private": true,
   "repository": {
     "type": "git",
@@ -12,91 +12,43 @@
     "url": "https://tice.software"
   },
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build && cp dist/en/index.html dist/",
-    "lint:vue": "vue-cli-service lint --no-fix",
-    "lint:tests": "eslint --no-fix test",
-    "lint": "yarn run lint:vue && yarn run lint:tests",
-    "test:local": "mocha test/local.test.js --timeout 60000",
-    "test:remote": "mocha test/remote.test.js",
-    "test:ci": "test/scripts/startServer.sh && yarn run test:local --reporter mocha-junit-reporter && test/scripts/stopServer.sh || test/scripts/stopServer.sh",
-    "docker": "docker build -t tice-web -f web.Dockerfile .",
-    "audit": "yarn audit --frozen-lockfile --level high --groups dependencies"
+    "dev": "vite",
+    "build": "run-p type-check build-only",
+    "preview": "vite preview",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --noEmit -p tsconfig.app.json --composite false",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore",
+    "format": "prettier --write src/"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.3.0",
-    "@fortawesome/free-brands-svg-icons": "^6.2.0",
-    "@fortawesome/vue-fontawesome": "^2.0.8",
     "core-js": "^3.26.0",
     "date-fns": "^2.29.3",
     "double-ratchet-ts": "^1.0.1",
-    "element-ui": "^2.15.10",
+    "element-plus": "^2.3.7",
     "elliptic": "^6.5.4",
     "jssha": "^3.3.0",
     "key-encoder": "^2.0.3",
     "libsodium-wrappers": "^0.7.10",
-    "mapbox-gl": "^2.10.0",
+    "pinia": "^2.1.4",
     "sodium-hkdf": "^1.0.0",
-    "vue": "^2.7.13",
-    "vue-beautiful-chat": "^2.5.0",
-    "vue-i18n": "^8.27.2",
-    "vue-mapbox": "^0.4.1",
-    "vue-timeago": "^5.1.3",
+    "vue": "^3.3.4",
+    "vue-i18n": "^9.2.2",
     "x3dh": "^1.0.0"
   },
   "devDependencies": {
-    "@types/libsodium-wrappers": "^0.7.10",
-    "@vue/cli-plugin-babel": "^4.5.19",
-    "@vue/cli-plugin-eslint": "^4.5.19",
-    "@vue/cli-service": "^4.5.19",
-    "@vue/eslint-config-airbnb": "^5.3.0",
-    "axios": "^0.27.2",
-    "babel-eslint": "^10.1.0",
-    "babel-plugin-component": "^1.1.1",
-    "chai": "^4.3.6",
-    "chromedriver": "^106.0.1",
-    "eslint": "^8.26.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-vue": "^9.6.0",
-    "geckodriver": "^3.2.0",
-    "html-webpack-plugin": "^4.5.2",
-    "mocha": "^10.1.0",
-    "mocha-junit-reporter": "^2.1.0",
-    "node-sass": "^7.0.3",
-    "sass-loader": "^13.1.0",
-    "selenium-webdriver": "^4.5.0",
-    "vue-template-compiler": "^2.7.13"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended",
-      "airbnb-base"
-    ],
-    "rules": {
-      "indent": [
-        "error",
-        4
-      ],
-      "max-len": "off",
-      "import/no-unresolved": "off",
-      "no-param-reassign": "off"
-    },
-    "parserOptions": {
-      "parser": "babel-eslint"
-    }
-  },
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions"
-  ]
+    "@rushstack/eslint-patch": "^1.2.0",
+    "@tsconfig/node18": "^2.0.1",
+    "@types/node": "^18.16.17",
+    "@vitejs/plugin-vue": "^4.2.3",
+    "@vue/eslint-config-prettier": "^7.1.0",
+    "@vue/eslint-config-typescript": "^11.0.3",
+    "@vue/tsconfig": "^0.4.0",
+    "eslint": "^8.39.0",
+    "eslint-plugin-vue": "^9.11.0",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.8.8",
+    "typescript": "~5.0.4",
+    "vite": "^4.3.9",
+    "vue-tsc": "^1.6.5"
+  }
 }
