+++ src/lang/index.js
@@ -1,0 +1,7 @@
+import en from "./en.json";
+import de from "./de.json";
+
+export const languages = {
+  en: en,
+  de: de
+};
